Activate is more user friendly, and there are a variety of graph structures to now choose from. All output is sent and compiled on designated PI. Currently no way to send anything to my Ubuntu subsystem on my Windows OS so that will have to do for now

+++ runratcon.sh
@@ -6,8 +6,27 @@ REMOTE='169.254.136.2'
 PINUMFILE='pinum.txt'
 PINUM=$(head -n 1 $PINUMFILE)
 FILENAME="output$PINUM.CSV"
+DEF_GRAPH="/home/pi/graphs/ring4.txt"
+GRAPH=$1
+REF_FILE=$2 
 
-`python ratio_consensus_dist.py`
-`cp output.CSV "$FILENAME"`
-`scp ./"$FILENAME" pi@"$REMOTE":~/csvFiles/`
-`ssh pi@"$REMOTE" "sh ~/runComp.sh"`
+if [[ -z "$GRAPH" ]]; then
+	GRAPH="$DEF_GRAPH"
+fi
+
+#getting rid of left over output file from previous run, if it exists
+if [[ -e "output.CSV" ]]; then
+	`rm output.CSV`
+fi
+
+echo "$REF_FILE"
+`python ratio_consensus_dist.py "$GRAPH" "$2"`
+#Checks if output.CSV exists first before sending
+#Send output to remote location and compile all into one file
+if [[ -e "output.CSV" ]]; then
+	`cp output.CSV "$FILENAME"`
+	`rm output.CSV`
+	`scp ./"$FILENAME" pi@"$REMOTE":~/csvFiles/`
+	`ssh pi@"$REMOTE" "sh ~/run_comp.sh $REF_FILE"`
+fi
+
